Backend/requirements.txt
# Backend/requirements.txt
fastapi
uvicorn[standard]

# Database
sqlalchemy
pydantic
python-dotenv

# Auth & Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2

# CORS
starlette
